# api_gateway/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SPOPA API Gateway",
    "description": "API Gateway for SPOPA Student Practices Management System",
    "version": "1.0.0",
    "contact": {
      "name": "SPOPA Team",
      "email": "support@spopa.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    },
    {
      "url": "https://api.spopa.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/test/load/{requests}": {
      "get": {
        "summary": "Load testing endpoint",
        "description": "Generates test responses for load testing (JMeter compatible)",
        "tags": ["Testing"],
        "parameters": [
          {
            "name": "requests",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test responses generated"
          }
        }
      }
    }
  },
...
}